Make placeholders more obvious
This commit changes the names of the {{ placeholders }} that the
package implementor has to fill in so that they are more obvious for
people who are committing their first beautifier.

They now all start with "beautifier". In addition, "package" has been
clarified as "NpmPackage". When I started out, the various "name"
placeholders caused me the most confusion, so they are now
consistently named "beautifierFancyTitle" and "beautifierDashedName".
So beautifierFancyTitle is the friendly human-readable name that is
often capitalized and may have spaces. beautifierDashedName is the
"foo-bar" part of the Unibeautify package name "beautifier-foo-bar".

Complete list of the new placeholder names:

{{ beautifierFancyTitle }}
{{ beautifierDashedName }}

{{ beautifierHomepageUrl }}
{{ beautifierInstallUrl }}
{{ beautifierBugsUrl }}

{{ beautifierDependencyType }}
{{ beautifierExeCommand }}
{{ beautifierNpmPackage }}